Fix PRNG seeding and multiaccuracy modifier application
- Fix PRNG random_chance_denom to use seed properly instead of always 0
- Fix multiaccuracy hits 2+ to use EventResult::Number for integer accuracy
  values so Gravity's onModifyAccuracy chain_modify is applied correctly
- Update various ability/item/move callbacks to use random_chance_denom
- Fix team generator PRNG seeding

+++ src/battle/random_chance.rs
@@ -9,7 +9,7 @@ impl Battle {
     // 		return this.prng.randomChance(numerator, denominator);
     // 	}
     //
-    pub fn random_chance(&mut self, numerator: i32, denominator: i32) -> bool {
+    pub fn random_chance(&mut self, numerator: f64, denominator: i32) -> bool {
         // JavaScript: if (this.forceRandomChance !== null) return this.forceRandomChance;
         // Check forced value FIRST before making PRNG call
         if let Some(forced) = self.force_random_chance {
